requirements.txt
# wiki_agent requirements
# Install with: pip install -r requirements.txt

requests>=2.31.0
rich>=13.0.0
python-docx>=1.1.0
pymupdf>=1.24.0        # PDF parsing (fitz) — preferred
pdfplumber>=0.10.0     # PDF parsing fallback
